Evaluate polynomials "normally"
Following Burden and Faires (and probably plently of others) we identify
the zero polynomial as having no defined degree, which we represent here
by null.  This allows us to not have to do anything fancy in CanExecute.

Also, we note the possibility of using System.Math.Pow instead of
repeated multiplication.  It could be that Pow accrues error differently
than the naïve approach.